app/ml-models/prescription-ocr/requirements.txt
# Prescription OCR - Complete Dependencies
# This file includes all dependencies for training, API service, and visualization
# Install with: pip install -r requirements.txt

# ============================================
# Core ML Frameworks (Training + API)
# ============================================
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# ============================================
# Image Processing (Training + API)
# ============================================
Pillow>=10.0.0
opencv-python>=4.8.0

# ============================================
# API Service Dependencies
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0

# ============================================
# Training & Dataset Dependencies
# ============================================
datasets>=2.12.0
evaluate>=0.4.0
albumentations>=1.3.0
imgaug>=0.4.0

# ============================================
# Data Processing & Utilities
# ============================================
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0

# ============================================
# Visualization & Charts
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================
# Jupyter Notebook Support (Optional)
# ============================================
# jupyter>=1.0.0
# ipykernel>=6.0.0

# Note: editdistance>=0.6.0 requires C++ compiler - optional for advanced metrics
